+++ packages/neon-core/__tests__/u/basic/base64.ts
@@ -1,4 +1,4 @@
-import { base642hex, hex2base64 } from "../../src/u/base64";
+import { base642hex, hex2base64 } from "../../src/u/basic/base64";
 
 const testCases = [
   ["4e454f", "TkVP"], // NEO
