+++ packages/neon-core/src/u/basic/random.ts
@@ -1,0 +1,9 @@
+import secureRandom from "secure-random";
+
+/**
+ * Generates a arrayBuffer filled with random bits.
+ * @param length - length of buffer.
+ */
+export const generateRandomArray = (length: number): number[] => {
+  return secureRandom(length) as number[];
+};
