--- packages/neon-core/src/u/random.ts
@@ -1,9 +1,0 @@
-import secureRandom from "secure-random";
-
-/**
- * Generates a arrayBuffer filled with random bits.
- * @param length - length of buffer.
- */
-export const generateRandomArray = (length: number): number[] => {
-  return secureRandom(length) as number[];
-};
